// clients/javascript/lib/types/CalendarEventDTO.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CalendarEventReminder } from "./CalendarEventReminder";
import type { ID } from "./ID";
import type { RRuleOptions } from "./RRuleOptions";

export type CalendarEventDTO = { id: ID, startTime: Date, duration: number, busy: boolean, updated: number, created: number, recurrence?: RRuleOptions, exdates: Array<Date>, calendarId: ID, userId: ID, reminders: Array<CalendarEventReminder>, metadata: Record<string, string>, };